+++ MoltenVK/MoltenVK/GPUObjects/MVKFramebuffer.h
@@ -21,6 +21,9 @@
 #include "MVKDevice.h"
 #include "MVKImage.h"
 #include "MVKSmallVector.h"
+#include <mutex>
+
+class MVKRenderSubpass;
 
 
 #pragma mark MVKFramebuffer
@@ -45,16 +48,25 @@ public:
 	/** Returns the attachments.  */
 	MVKArrayRef<MVKImageView*> getAttachments() { return _attachments.contents(); }
 
+	/**
+	 * Returns a MTLTexture for use as a dummy texture when a render subpass,
+	 * that is compatible with the specified subpass, has no attachments.
+	 */
+	id<MTLTexture> getDummyAttachmentMTLTexture(MVKRenderSubpass* subpass, uint32_t passIdx);
+
 #pragma mark Construction
 
-	/** Constructs an instance for the specified device. */
 	MVKFramebuffer(MVKDevice* device, const VkFramebufferCreateInfo* pCreateInfo);
+
+	~MVKFramebuffer() override;
 
 protected:
 	void propagateDebugName() override {}
 
+	MVKSmallVector<MVKImageView*, 4> _attachments;
+	id<MTLTexture> _mtlDummyTex = nil;
+	std::mutex _lock;
 	VkExtent2D _extent;
 	uint32_t _layerCount;
-	MVKSmallVector<MVKImageView*, 4> _attachments;
 };